+++ package.json
@@ -12,7 +12,6 @@
     "lint": "eslint . --ext ts --ext tsx --max-warnings 0"
   },
   "dependencies": {
-    "@radix-ui/react-dialog": "^1.0.4",
     "@types/node": "20.3.0",
     "@types/react": "18.2.11",
     "@types/react-dom": "18.2.4",
@@ -26,17 +25,17 @@
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-spring": "^9.7.1",
     "surrealdb.js": "^0.8.0",
     "swr": "^2.1.5",
     "tailwind-merge": "^1.13.1",
     "tailwindcss": "3.3.2",
-    "typescript": "5.1.3"
+    "typescript": "5.1.3",
+    "zustand": "^4.3.8"
   },
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.59.9",
+    "@typescript-eslint/parser": "^5.59.9",
     "concurrently": "^8.2.0",
-    "prettier-plugin-organize-imports": "^3.2.2",
-    "prettier-plugin-tailwindcss": "^0.3.0",
     "eslint-config-next": "13.4.5",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-json": "^3.1.0",
@@ -45,7 +44,7 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-unused-imports": "^2.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9"
+    "prettier-plugin-organize-imports": "^3.2.2",
+    "prettier-plugin-tailwindcss": "^0.3.0"
   }
 }
