+++ src/app/page.tsx
@@ -1,12 +1,26 @@
 'use client';
 
-import { Sticky } from '@/components/sticky';
-import { useStickies } from '@/lib/hooks';
+import { Sticky, StickyFramework } from '@/components/sticky';
+import { useCreateSticky, useStickies } from '@/lib/hooks';
+import { useCreateStickyStore } from '@/lib/state';
 import { Loader2 } from 'lucide-react';
-import React from 'react';
+import React, { useCallback } from 'react';
 
 export default function Home() {
     const { data, error, isLoading } = useStickies();
+    const { createColor, setCreateColor } = useCreateStickyStore((s) => s);
+    const { trigger: createSticky } = useCreateSticky();
+
+    const submit = useCallback(
+        (content: string) => {
+            if (createColor) {
+                createSticky({ color: createColor, content });
+                setCreateColor(null);
+            }
+        },
+        [createColor, createSticky, setCreateColor]
+    );
+
     const message =
         data?.stickies.length == 0 ? (
             'Create a sticky!'
@@ -29,6 +43,15 @@ export default function Home() {
         </div>
     ) : (
         <div className="grid w-full grid-cols-1 gap-6 sm:py-16 md:grid-cols-2 xl:grid-cols-3">
+            {createColor && (
+                <StickyFramework
+                    color={createColor}
+                    onClose={() => setCreateColor(null)}
+                    onDelete={() => setCreateColor(null)}
+                    onSubmit={submit}
+                    editing={true}
+                />
+            )}
             {data?.stickies.map(({ id, content, color }) => (
                 <Sticky key={id} id={id} color={color} content={content} />
             ))}
