+++ src/components/sticky.tsx
@@ -1,27 +1,11 @@
 'use client';
 
 import { useDeleteSticky, useUpdateSticky } from '@/lib/hooks';
+import { StickyColor, style } from '@/lib/style';
 import { cn } from '@/lib/utils';
-import { cva } from 'class-variance-authority';
 import { X } from 'lucide-react';
 import React, { createRef, useCallback, useEffect, useState } from 'react';
-import { Textarea } from './sticky-editor';
-
-const style = cva('', {
-    variants: {
-        color: {
-            purple: 'bg-purple-sticky',
-            pink: 'bg-pink-sticky',
-        },
-    },
-});
-
-// Looks more complex than it actually is
-// This essentially extracts a type for the colors that we can pass
-export type StickyColor = Exclude<
-    Exclude<Parameters<typeof style>[0], undefined>['color'],
-    null | undefined
->;
+import { Textarea } from './textarea';
 
 export function Sticky({
     id,
@@ -32,7 +16,6 @@ export function Sticky({
     color: StickyColor;
     content: string;
 }) {
-    const ref = createRef<HTMLDivElement>();
     const [editing, setEditing] = useState(false);
     const { trigger: deleteSticky } = useDeleteSticky(id);
     const { trigger: updateSticky } = useUpdateSticky(id);
@@ -45,15 +28,47 @@ export function Sticky({
         [color, updateSticky]
     );
 
+    return (
+        <StickyFramework
+            color={color}
+            content={content}
+            onClick={() => setEditing(true)}
+            onClose={() => setEditing(false)}
+            onDelete={deleteSticky}
+            onSubmit={submit}
+            editing={editing}
+        />
+    );
+}
+
+export function StickyFramework({
+    color,
+    content,
+    onClick,
+    onClose,
+    onSubmit,
+    onDelete,
+    editing,
+}: {
+    color: StickyColor;
+    content?: string;
+    onClick?: () => unknown;
+    onClose?: () => unknown;
+    onSubmit?: (content: string) => unknown;
+    onDelete?: () => unknown;
+    editing?: boolean;
+}) {
+    const ref = createRef<HTMLDivElement>();
+
     useEffect(() => {
         const handler = (event: globalThis.MouseEvent) => {
             if (ref.current && !ref.current.contains(event.target as Node)) {
-                setEditing(false);
+                onClose?.();
             }
         };
         window.addEventListener('mousedown', handler);
         return () => window.removeEventListener('mousedown', handler);
-    }, [ref]);
+    }, [ref, onClose]);
 
     return (
         <div
@@ -65,7 +80,7 @@ export function Sticky({
         >
             <button
                 className="absolute right-0 top-0 z-10 m-4 border-none bg-transparent p-0 text-white outline-none"
-                onClick={() => deleteSticky()}
+                onClick={onDelete}
             >
                 <X size={30} />
             </button>
@@ -74,14 +89,14 @@ export function Sticky({
                     'relative flex aspect-square w-full cursor-pointer flex-col items-start rounded-4xl border-none px-10 py-14 text-left text-2xl text-dark outline-none',
                     style({ color })
                 )}
-                onClick={() => setEditing(true)}
+                onClick={onClick}
             >
                 {editing ? (
                     <Textarea
                         initialValue={content}
-                        onSubmit={submit}
+                        onSubmit={onSubmit}
                         editing={editing}
-                        setEditing={setEditing}
+                        setEditing={(s) => (s ? onClick : onClose)?.()}
                     />
                 ) : (
                     <p>{content}</p>
